+++ packages/instantsearch.js/src/widgets/hits/defaultTemplates.ts
@@ -1,7 +1,7 @@
 import { omit } from '../../lib/utils';
 
 // false positive lint error
-// eslint-disable-next-line @typescript-eslint/consistent-type-imports
+// eslint-disable-next-line typescript/consistent-type-imports
 import type { HitsTemplates } from './hits';
 
 const defaultTemplates = {