+++ packages/react-instantsearch-core/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.28.0](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.27.1...react-instantsearch-core@7.28.0) (2026-03-16)
+
+**Note:** Version bump only for package react-instantsearch-core
+
+
+
+## [7.27.1](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.27.0...react-instantsearch-core@7.27.1) (2026-03-16)
+
+**Note:** Version bump only for package react-instantsearch-core
+
+
+
+
+
 ## [7.27.0](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.26.1...react-instantsearch-core@7.27.0) (2026-03-10)